safari-600.7-branch/LayoutTests/fast/speech/scripted/navigate-away-expected.txt

Test behaviour when navigating away from a page using the Speech JavaScript API

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


The iframe was loaded.
PASS 'webkitSpeechRecognition' in self is true
PASS webkitSpeechRecognition == null is false
PASS testRunner.wasMockSpeechRecognitionAborted() is false
iframe: Created SpeechRecognition
iframe: calling start()
iframe: onstart
iframe: navigating away
The iframe was navigated away.
PASS testRunner.wasMockSpeechRecognitionAborted() is true
PASS successfullyParsed is true

TEST COMPLETE
